inital commit
code the Cline AI made for me with this specific prompt

I run a 389DS ldap server and i want to make a script that checks for replication agreements on suppliers and hub and then updates the supplier and hub agreements passwords along with the passwords on al the consumers hooked to these agreements. Act as a RHEL7/RHEL8 389DS server admin who is learning go and want to build an application to help manage these passwords, you are also familiar with error code 49 and are very intrested in GRPC and how it can be used to catch these errors on the fly. This code should be able to be tested and print out what it would change the passwords too on what servers it would opearte on and print out a LDAP command todo that task by hand if desired. Please give back the soultion you come up with in the form of files and code. The code must be commented with at least 5 lines that explain what is going on, why and how. The program should be understantable by a non programmer and able to be updated by a non developer as well and used to teach people how todo it. Keep all the code following the Kiss pincipal and easy to follow and understand. explain each though of the code as you go. an exmaple would in the /var/log/dirsrv/slapd-ldap/errors the log will report error:49 followed by the name of the replication agreement and this means the passwords on either the supplier or consumer need to be updated